Change license from MIT to BUSL-1.1 (#385)
I have put a lot of thought into this and decided to make this change to
protect this project from outside sources. As a result, I am purposing
of changing the license from MIT to BUSL-1.1.

For nearly all possible use cases, I want to make it very clear, this
project is **still** MIT. Anyone that wants to make derivative works or
use the project may do so.

The license is being changed to BUSL to prevent the two following use
cases:

### Ubiquiti may not use this library in any capacity

Internally for development, for tooling, or for documentations or
recommendations to users. Ubiquiti has been increasingly hostile towards
local only access. They are so insistent on trying to create a facade
they support "local only", but then lock more things behind Remote
Access and their cloud. As long as they want to try to gaslight the
community and keep telling them "just enable Remote Access" to access
local only features, they cannot use this library in any capacity. Just
in case they happen to already use it.

### This library may not be used for Home Assistant except if it comes
_from this repo_

It may not be forked and published to be used in HA. Nabu Casa employees
have made [threats to try to cut this project out of Home
Assistant](https://github.com/AngellusMortis/pyunifiprotect/pull/243#issuecomment-1281627815)
just because they disagree with how software is written for this
project. They have also tried to force people and projects to [not allow
the usage of their software in other
projects](https://github.com/NixOS/nixpkgs/pull/126326#issuecomment-860047453)
as well. The more I think about the more that is really not okay. If
Home Assistant does not like how `unifiprotect` is managed, they can
kick me out of being a code owner, but they may not benefit from my
efforts of building a Python API for UniFi Protect as a result. If they
want to micromanage every single aspect of code in Home Assistant, they
can build their own Python API for UniFi Protect. Nabu Casa always
speaks to how they support open source, but then they allow their
employees to try to bully other projects into following their rules and
their standards.

I know BUSL and other "source available" licenses are often very
controversial because it goes against the absolutist nature of OSI, but
OSI has a really short-sighted view on how the world works. In a perfect
world, OSI licenses should work. There is no reason MIT should not "just
work" for everyone. But the reality is that larger groups will always
use their influence and control to bully smaller groups.

To show that I mean what I say, the Change Date for BUSL is only 6
months, which is much lower than the required 4 years. This means if for
whatever reason, this library stops being maintained, it can quickly be
forked and transferred to a new owner if they choose to.