Remove global recording settings from cameras (#393)
Protect 4.0 seems to have deprecated global recording settings. They are
no longer in the UI and are replaced with the "Recording Manager".
Recording Manager lets you bulk edit camera settings. Any edits with
Recording Manager seems to disable `use_global` on the camera and there
is no way via the UI to turn it back on.


This removes checking `use_global` from everywhere and anywhere that was
raising an exception before because it was changing a recording setting
now does the same thing as the Recording Manager does and turns _off_
the `use_global` flag first before editing.